+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Fixes
+
+- Fixed incorrect amount of token sent ([PR 2962](https://github.com/input-output-hk/daedalus/pull/2962))
+
 ### Chores
 
 - Added Vasil-supported cardano-wallet ([PR 3001](https://github.com/input-output-hk/daedalus/pull/3001))